django_celery_boilerplate/celery.py
import os

from celery import Celery
from dotenv import find_dotenv, load_dotenv

load_dotenv(find_dotenv())

APP_NAME = os.getenv("APP_NAME")

# set the default Django settings module for the 'celery' program.
os.environ.setdefault("DJANGO_SETTINGS_MODULE", f"{APP_NAME}.settings")

app = Celery(APP_NAME)

# Using a string here means the worker doesn't have to serialize
# the configuration object to child processes.
# - namespace='CELERY' means all celery-related configuration keys
#   should have a `CELERY_` prefix.
app.config_from_object("django.conf:settings", namespace="CELERY")

# Load task modules from all registered Django app configs.
app.autodiscover_tasks()


@app.task(bind=True)
def debug_task(self):
    print(f"Request: {self.request!r}")
