pyproject.toml
[tool.poetry]
name = "django_celery_boilerplate"
version = "0.1.0"
description = ""
authors = ["Evgeny Kemerov <eskemerov@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
python-dotenv = "^0.14.0"
psycopg2-binary = "^2.8"
sentry-sdk = "^0.16.0"
django = "^3.0"
djangorestframework = "^3.11.0"

[tool.poetry.dev-dependencies]
pylint = "^2.3"
flake8 = "^3.7"
isort = "^5.0"
pylint-django = "^2.0"
black = "^19.10b0"
ipython = "^7.9.0"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
